add restriction for adi activities for next/previous functionality
