+++ src/app/pages/candidate-details/__tests__/candidate-details.page.spec.ts
@@ -30,6 +30,9 @@ import { SlotItem } from '@providers/slot-selector/slot-item';
 import { JournalModel } from '@store/journal/journal.model';
 import { StoreModel } from '@shared/models/store.model';
 import { TestStatus } from '@store/tests/test-status/test-status.model';
+import { SlotProvider } from '@providers/slot/slot';
+import { AppConfigProvider } from '@providers/app-config/app-config';
+import { AppConfigProviderMock } from '@providers/app-config/__mocks__/app-config.mock';
 import { CandidateDetailsPage } from '../candidate-details.page';
 
 describe('CandidateDetailsPage', () => {
@@ -97,6 +100,8 @@ describe('CandidateDetailsPage', () => {
         { provide: ModalController, useClass: ModalControllerMock },
         { provide: NavParams, useValue: mockNavParams },
         { provide: Router, useClass: RouterMock },
+        { provide: SlotProvider, useClass: SlotProvider },
+        { provide: AppConfigProvider, useClass: AppConfigProviderMock },
         provideMockStore({ initialState }),
       ],
     });
